pyproject.toml
[tool.poetry]
name = "before-you-go"
version = "0.1.3"
description = ""
authors = ["Lester <l.ester@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
Jinja2 = "^3.1.4"
SQLAlchemy = "^2.0.31"
uvicorn = "^0.30.3"
sqlmodel = "^0.0.21"
bcrypt = "^4.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.32.3"
openfoodfacts = "^1.1.1"
pillow = "^10.4.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
alembic = "^1.13.2"
starlette-session-middleware = "^0.1.1"
itsdangerous = "^2.2.0"
resend = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
